feat: implement BranchPicker component for selecting git branches
