+++ src/vs/sessions/contrib/chat/browser/sessionTargetPicker.ts
@@ -14,8 +14,6 @@ import { Disposable, DisposableStore } from '../../../../base/common/lifecycle.j
 import { renderIcon } from '../../../../base/browser/ui/iconLabel/iconLabels.js';
 import { localize } from '../../../../nls.js';
 import { IContextMenuService } from '../../../../platform/contextview/browser/contextView.js';
-import { IActionWidgetService } from '../../../../platform/actionWidget/browser/actionWidget.js';
-import { ActionListItemKind, IActionListDelegate, IActionListItem } from '../../../../platform/actionWidget/browser/actionList.js';
 import { AgentSessionProviders } from '../../../../workbench/contrib/chat/browser/agentSessions/agentSessions.js';
 import { IGitRepository } from '../../../../workbench/contrib/git/common/gitService.js';
 import { INewSession } from './newSession.js';
@@ -277,202 +275,6 @@ export class IsolationModePicker extends Disposable {
 			this._onDidChange.fire(mode);
 			this._renderDropdown();
 		}
-	}
-}
-
-// #endregion
-
-// #region --- Branch Picker ---
-
-const COPILOT_WORKTREE_PREFIX = 'copilot-worktree-';
-
-interface IBranchItem {
-	readonly name: string;
-}
-
-/**
- * A self-contained widget for selecting a git branch.
- * Uses `IGitRepository.getRefs` to list local branches, grouped into
- * "Copilot Worktrees" and "Branches" sections.
- * Writes the selected branch to the new session object.
- */
-export class BranchPicker extends Disposable {
-
-	private _selectedBranch: string | undefined;
-	private _newSession: INewSession | undefined;
-	private _repository: IGitRepository | undefined;
-	private _branches: string[] = [];
-
-	private readonly _onDidChange = this._register(new Emitter<string | undefined>());
-	readonly onDidChange: Event<string | undefined> = this._onDidChange.event;
-
-	private readonly _renderDisposables = this._register(new DisposableStore());
-	private _container: HTMLElement | undefined;
-	private _triggerElement: HTMLElement | undefined;
-
-	get selectedBranch(): string | undefined {
-		return this._selectedBranch;
-	}
-
-	constructor(
-		@IActionWidgetService private readonly actionWidgetService: IActionWidgetService,
-	) {
-		super();
-	}
-
-	/**
-	 * Sets the new session that this picker writes to.
-	 */
-	setNewSession(session: INewSession | undefined): void {
-		this._newSession = session;
-	}
-
-	/**
-	 * Sets the git repository and loads its branches.
-	 * When undefined, the picker is hidden.
-	 */
-	async setRepository(repository: IGitRepository | undefined): Promise<void> {
-		this._repository = repository;
-		this._branches = [];
-		this._selectedBranch = undefined;
-
-		if (repository) {
-			const refs = await repository.getRefs({ pattern: 'refs/heads' });
-			this._branches = refs
-				.map(ref => ref.name)
-				.filter((name): name is string => !!name);
-		}
-
-		this.setVisible(!!repository && this._branches.length > 0);
-		this._updateTriggerLabel();
-	}
-
-	/**
-	 * Renders the branch picker trigger into the given container.
-	 */
-	render(container: HTMLElement): void {
-		this._container = container;
-		this._renderDisposables.clear();
-
-		const trigger = dom.append(container, dom.$('a.action-label'));
-		trigger.tabIndex = 0;
-		trigger.role = 'button';
-		this._triggerElement = trigger;
-		this._updateTriggerLabel();
-
-		this._renderDisposables.add(dom.addDisposableListener(trigger, dom.EventType.CLICK, (e) => {
-			dom.EventHelper.stop(e, true);
-			this.showPicker();
-		}));
-
-		this._renderDisposables.add(dom.addDisposableListener(trigger, dom.EventType.KEY_DOWN, (e) => {
-			if (e.key === 'Enter' || e.key === ' ') {
-				dom.EventHelper.stop(e, true);
-				this.showPicker();
-			}
-		}));
-	}
-
-	/**
-	 * Shows or hides the picker.
-	 */
-	setVisible(visible: boolean): void {
-		if (this._container) {
-			this._container.style.visibility = visible ? '' : 'hidden';
-		}
-	}
-
-	/**
-	 * Shows the branch picker dropdown anchored to the trigger element.
-	 */
-	showPicker(): void {
-		if (!this._triggerElement || this.actionWidgetService.isVisible) {
-			return;
-		}
-
-		const items = this._buildItems();
-		const triggerElement = this._triggerElement;
-		const delegate: IActionListDelegate<IBranchItem> = {
-			onSelect: (item) => {
-				this.actionWidgetService.hide();
-				this._selectBranch(item.name);
-			},
-			onHide: () => { triggerElement.focus(); },
-		};
-
-		this.actionWidgetService.show<IBranchItem>(
-			'branchPicker',
-			false,
-			items,
-			delegate,
-			this._triggerElement,
-			undefined,
-			[],
-			{
-				getAriaLabel: (item) => item.name,
-				getWidgetAriaLabel: () => localize('branchPicker.ariaLabel', "Branch Picker"),
-			},
-			this._branches.length > 10 ? { showFilter: true, filterPlaceholder: localize('branchPicker.filter', "Filter branches...") } : undefined,
-		);
-	}
-
-	private _buildItems(): IActionListItem<IBranchItem>[] {
-		const items: IActionListItem<IBranchItem>[] = [];
-		const worktreeBranches = this._branches.filter(b => b.startsWith(COPILOT_WORKTREE_PREFIX));
-		const otherBranches = this._branches.filter(b => !b.startsWith(COPILOT_WORKTREE_PREFIX));
-
-		if (worktreeBranches.length > 0) {
-			items.push({
-				kind: ActionListItemKind.Header,
-				label: localize('branchPicker.worktrees', "Copilot Worktrees"),
-			});
-			for (const branch of worktreeBranches) {
-				items.push({
-					kind: ActionListItemKind.Action,
-					label: branch,
-					group: { title: '', icon: this._selectedBranch === branch ? Codicon.check : Codicon.blank },
-					item: { name: branch },
-				});
-			}
-		}
-
-		if (otherBranches.length > 0) {
-			items.push({
-				kind: ActionListItemKind.Header,
-				label: localize('branchPicker.branches', "Branches"),
-			});
-			for (const branch of otherBranches) {
-				items.push({
-					kind: ActionListItemKind.Action,
-					label: branch,
-					group: { title: '', icon: this._selectedBranch === branch ? Codicon.check : Codicon.blank },
-					item: { name: branch },
-				});
-			}
-		}
-
-		return items;
-	}
-
-	private _selectBranch(branch: string): void {
-		if (this._selectedBranch !== branch) {
-			this._selectedBranch = branch;
-			this._newSession?.setBranch(branch);
-			this._onDidChange.fire(branch);
-			this._updateTriggerLabel();
-		}
-	}
-
-	private _updateTriggerLabel(): void {
-		if (!this._triggerElement) {
-			return;
-		}
-		dom.clearNode(this._triggerElement);
-		const label = this._selectedBranch ?? localize('branchPicker.select', "Branch");
-		dom.append(this._triggerElement, renderIcon(Codicon.gitBranch));
-		const labelSpan = dom.append(this._triggerElement, dom.$('span.sessions-chat-dropdown-label'));
-		labelSpan.textContent = label;
-		dom.append(this._triggerElement, renderIcon(Codicon.chevronDown));
 	}
 }
